Factor out slate utils into own files
