+++ client/src/modules/blockEditor/actions.ts
@@ -1,7 +1,7 @@
 import * as uuidv1 from "uuid/v1";
 import { UPDATE_BLOCK } from "../blocks/actions";
 import { SPACER } from "../../lib/slate-pointers/exportedPointerSpacer";
-import { normalizeAfterRemoval } from "./normalizeAfterRemoval";
+import { normalizeAfterRemoval } from "../../utils/slate/normalizeAfterRemoval";
 
 export const CHANGE_HOVERED_ITEM = "CHANGE_HOVERED_ITEM";
 export const CHANGE_POINTER_REFERENCE = "CHANGE_POINTER_REFERENCE";
